feat: 홈 화면 및 지도 컴포넌트 분리
